Interim proof-of-concept commit for dynamic report buttons.
Unique array of upchain and downchain processor types is populated and
passed to ui.  the ui now inspects the selectedTimeSeries object to
determine with report to show.  I'm concerned about race conditions,
though, if the user tries to open another "folder" while another set of
report buttons are still rendering.